requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
argilla==1.12.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
configobj==5.0.8
configparser==6.0.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.100.0
fastjsonschema==2.17.1
ffmpy==0.3.0
filelock==3.12.2
filetype==1.2.0
fitz==0.0.1.dev2
Flask==2.3.2
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
google-search-results==2.4.2
gradio==3.36.1
gradio_client==0.2.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
langchain==0.0.225
langchainplus-sdk==0.0.20
linkify-it-py==2.0.2
looseversion==1.3.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pdf2image==1.16.3
pdfminer.six==20221105
pgvector==0.1.8
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
prov==2.0.0
psutil==5.9.5
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.11
pydot==1.4.2
pydub==0.25.1
Pygments==2.15.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.12.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pywin32==306
pywinpty==2.0.10
pyxnat==1.5
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rdflib==6.3.2
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rpds-py==0.8.8
scipy==1.11.1
semantic-version==2.10.0
Send2Trash==1.8.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.3.2
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
unstructured==0.8.0
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wikipedia==1.4.0
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
